% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BuildExPostRtdLoss}
\alias{BuildExPostRtdLoss}
\title{Build loss function of rotation matrix}
\usage{
BuildExPostRtdLoss(Lambda, LambdaStar, PhiBar, PhiStar)
}
\arguments{
\item{Lambda}{loadings matrix.}

\item{LambdaStar}{objective loadings matrix.}

\item{PhiBar}{matrix of VAR parameters.}

\item{PhiStar}{objetive matrix of VAR parameters.}
}
\value{
A function with two arguments:
\itemize{
  \item \code{parms} (which contains the parameters that build the orthogonal
    matrix); and
  \item \code{reflexion} (element which defines if the orthogonal matrix has
    positive or negative determinant).
}
}
\description{
This function builds the quadratic loss function of a
  rotation matrix using the \code{BuildOrthMat} function parameterisation.
}

