% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BuildOrthMat}
\alias{BuildOrthMat}
\title{Build orthogonal matrix}
\usage{
BuildOrthMat(parms)
}
\arguments{
\item{parms}{numeric vector of rotation angles.}
}
\value{
Orthogonal matrix.
}
\description{
Build orthogonal matrix from the formula 24 presented in Assmann et al. (2014).
}
\references{
Assmann C, Boysen-Hogrefe J and Pape M (2014). "Bayesian Analysis of Dynamic Factor
Models: An Ex-Post Approach towards the Rotation Problem." \emph{Technical Report 1902}, Kiel
Institution for the World Economy.
}

