% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExPostLoss}
\alias{ExPostLoss}
\title{Ex-post quadratic loss}
\usage{
ExPostLoss(Lambda, LambdaStar, PhiBar, PhiStar)
}
\arguments{
\item{Lambda}{DFM loadings matrices piled in a single matrix.}

\item{LambdaStar}{objective piled loadings matrix.}

\item{PhiBar}{matrix of VAR parameters joined side by side.}

\item{PhiStar}{objetive joined matrices of VAR parameters.}
}
\value{
The quadratic loss is returned.
}
\description{
Function that calculates the quadratic loss function of the
  ex-post approach to solve the rotation problem.
}

