% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExPostLossFunction}
\alias{ExPostLossFunction}
\title{Ex-post quadratic loss function.}
\usage{
ExPostLossFunction(Lambda, parms, LambdaStar, PhiBar, PhiStar)
}
\arguments{
\item{Lambda}{matrix of loadings of contemporaneous and lagged factors.
See details.}

\item{parms}{vector of rotation angles to build the orthogonal matrix.}

\item{LambdaStar}{matrix of loadings to be achieved by a loss minimization
procedure.}

\item{PhiBar}{matrix concatenation of VAR parameter matrices. See details.}

\item{PhiStar}{VAR parameter to be achieved by a procedure of loss
minimization.}
}
\value{
Quadratic loss and, as an attribute, the best sign of the orthogonal
matrix.
}
\description{
This function returns the quadratic loss function
of an orthogonal transformation of the loadings and the VAR parameters
in a DFM (\emph{dynamic factor model}).
}
\details{
The arguments \code{Lambda} and \code{LambdaStar} must be
a row binding of all loadings matrices. For example, suppose we have
\eqn{s = 2}. In this case, we have three matrices of loadings which are
\eqn{\Lambda_0}{\Lambda[0]}, \eqn{\Lambda_1}{\Lambda[1]} and
\eqn{\Lambda_2}{\Lambda[2]}. So, \code{Lambda} is equal to
\eqn{(\Lambda_0, \Lambda_1', \Lambda_2')'}{(\Lambda[0]^T,\Lambda[1]^T,\Lambda[2]^T)^T}.
}

