% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExPostLossOptim}
\alias{ExPostLossOptim}
\title{DFM quadratic loss optimisation}
\usage{
ExPostLossOptim(Lambda, LambdaStar, PhiBar, PhiStar)
}
\arguments{
\item{Lambda}{loadings matrix. See details.}

\item{LambdaStar}{loadings matrix objective.}

\item{PhiBar}{VAR parameters defining the factors evolution.}

\item{PhiStar}{objective value of PhiBar.}
}
\value{
A list which contains rotation matrix, rotated loadings,
and rotated PhiBar.
}
\description{
This function enables one to optimise the quadratic loss
of a DFM.
}

