% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fdlmGibbsIter}
\alias{fdlmGibbsIter}
\title{fdlm - iteracao do Gibbs}
\usage{
fdlmGibbsIter(model)
}
\arguments{
\item{model}{objeto tipo 'list' que contem os dados de entrada, as variaveis exogenas
e os hiperparametros das prioris.}
}
\value{
function cujo o parametro e' um valor logico indicando se e para retornar
os valores da cadeia ou somente avancar nas iteracoes.
}
\description{
Funcao a ser utilizada para se ter o resultado de uma iteracao do algoritmo de Gibbs.
Com seu uso e' possivel aplicar o modelo fdlm a muitas series a alocar as amostras da
posteriori a um objeto 'ff'.
}
\details{
Nessa funcao e' necessario especificar o objeto \code{model} com os
seguintes nomes:
\itemize{
\item \code{y}: matriz de dados (tempo x nVars);
\item \code{xFixReg}: matriz de variaveis explicativas da regressao estatica;
\item \code{xDynReg}: matriz de variaveis explicativas da regressao dinamica;
\item \code{nFactors}: numero de fatores latentes;
\item \code{L0}: media \emph{a priori} da matriz (nVars x nFactors) de cargas;
\item \code{H0}: matriz (nFactors x nFactors) de covariancias das
linhas da matriz de cargas;
\item \code{m0}: media \emph{a priori} do vetor (nVars*nxDyn x 1) dos
parametros de estado;
\item \code{C0}: variancia inicial dos parametros de estado;
\item \code{b0}: matriz de media dos parametros da regressao estatica;
\item \code{B0}: matriz (nxFix x nxFix) de covariancias dos parametros da regressao estatica;
\item \code{n0}: graus de liberdade \emph{a priori} das variancias idiossincraticas;
\item \code{s0sq}: vetor (nVars x 1) de medias harmonicas das variancias idiossincratica;
\item \code{discW}: fator de desconto dos parametros de estado.
}
}

