% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fdlmGibbsNA}
\alias{fdlmGibbsNA}
\title{Algoritmo de Gibbs para F-DLM com NA}
\usage{
fdlmGibbsNA(N, brn, thn, model, initVal, progressBar = TRUE,
  onlyValues = FALSE)
}
\arguments{
\item{N}{tamanho final da amostra.}

\item{brn}{periodo de aquecimento do algoritmo.}

\item{thn}{espacamento entre as extracoes.}

\item{model}{modelo F-DLM com valores da matriz de dados, de exogenas e hiperparametros da priori.}

\item{initVal}{valores iniciais para Beta, Lambda e psi.}
}
\value{
Lista com modelo e valores simulados do Gibbs.
}
\description{
Estima as quantidades do modelo dinamico fatorial de acordo com os resultados presentes na dissertacao.
}

