% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetLogX2ApproxParms}
\alias{GetLogX2ApproxParms}
\title{Mixture of normals approximation}
\usage{
GetLogX2ApproxParms(y)
}
\arguments{
\item{y}{data matrix that follows a log-\eqn{\chi^2} distribution. The matrix
must be entered with variables in columns.}
}
\value{
A list containing the mean and the variances parameters of the normal
  distributions.
}
\description{
Get normal parameters of the gaussian distributions which
  approximates the log-\eqn{\chi^2} distribution. This approximation is made
  considering 10 gaussian kernels such as those presented in Omori
  \emph{et al.} (2007).
}

