% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mdfDiscW.sim}
\alias{mdfDiscW.sim}
\title{Simula F-DLM}
\usage{
mdfDiscW.sim(xFixReg, parmsFixReg, xDynReg, m0, C0, discW = 0.95, Lambda, psi,
  seed = as.numeric(Sys.time()))
}
\arguments{
\item{xFixReg}{matriz de delinameanto da regressao estatica;}

\item{parmsFixReg}{matriz de parametros da regressao estatica;}

\item{xDynReg}{matriz de delineamento da regressao dinamica;}

\item{m0,C0}{media e variancia \emph{a priori} dos estados;}

\item{discW}{fator de desconto;}

\item{Lambda}{matriz de cargas fatoriais;}

\item{psi}{vetor de variancias idiossincraticas;}

\item{seed}{semente do gerador aleatorio.}
}
\value{
Lista com os parametros de entrada, os parametros simulados e os dados artificiais.
}
\description{
Funcao que simula dados de um modelo dinamico fatorial
}

