% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plt.fdlm}
\alias{plt.fdlm}
\title{Matriz de cargas via PLT}
\usage{
plt.fdlm(Lambda)
}
\arguments{
\item{Lambda}{array com as simulacoes da matriz de cargas.}
}
\value{
Lista com array de cargas sujeitos a PLT e matriz ortogonal da restricao.
}
\description{
Aplica restricao PLT na matriz de cargas do MCMC.
}

