% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RotatePhi}
\alias{RotatePhi}
\title{Rotate matrix of VAR parameters}
\usage{
RotatePhi(PhiBar, D)
}
\arguments{
\item{PhiBar}{matrix of binded matrices of VAR parameters.}

\item{D}{rotation matrix.}
}
\value{
Rotated \code{PhiBar}.
}
\description{
Given an orthogonal matrix D, do right and left rotations to
  the matrix of VAR parameters.
}
\details{
The rotation is of the form \eqn{D' \Phi[j] D}, where
\eqn{j = 1, \ldots , h}.
}

