% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RunDfmExPostOptAlg}
\alias{RunDfmExPostOptAlg}
\title{Ex-post DFM loss optimisation}
\usage{
RunDfmExPostOptAlg(LambdaBar, PhiBar, tol = 1e-07, max.iter = 10)
}
\arguments{
\item{LambdaBar}{array of simulated loadings matrix.}

\item{PhiBar}{array of simulated VAR parameters.}

\item{tol}{tolerance level of the optimisation.}

\item{max.iter}{maximum number of iterations for the algorithm.}
}
\value{
A list containing the rotated parameters as well as the objective
  (best) values for them.
}
\description{
Apply the optimisation algorithm that solves the rotation
  problem for the dynamic factor model.
}

