% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RunFdlmIntGibbs}
\alias{RunFdlmIntGibbs}
\title{Amostrador de Gibbs para F-DRM com intercepto}
\usage{
RunFdlmIntGibbs(N, brn, thn, model, initVal, progressBar = TRUE,
  onlyValues = FALSE)
}
\arguments{
\item{N}{tamanho final da amostra.}

\item{brn}{periodo de aquecimento do algoritmo.}

\item{thn}{espacamento entre as extracoes.}

\item{model}{modelo F-DLM com valores da matriz de dados,
de exogenas e hiperparametros da priori.}

\item{initVal}{valores iniciais para Beta, Lambda e psi.}
}
\value{
Lista com modelo e valores simulados do Gibbs.
}
\description{
Roda o amostrador de Gibbs para F-DRM onde o intercepto
 e os parametros de estados sao amostrados conjuntamente.
}

