% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wop.fdlm}
\alias{wop.fdlm}
\title{Matriz de cargas via WOP}
\usage{
wop.fdlm(Lambda, max.iter = 100)
}
\arguments{
\item{Lambda}{array oriundo do MCMC;}

\item{max.iter}{maximo de iteracoes ate convergencia.}
}
\value{
Lista com matriz ortogonal \code{D} e cargas tendo aplicado o WOP.
}
\description{
Aplica metodo WOP para garantir unicidade da matriz de cargas.
}

