% Generated by roxygen2 (4.0.1): do not edit by hand
\name{WopDynFactors}
\alias{WopDynFactors}
\title{Ex-post quadratic loss optimization}
\usage{
WopDynFactors(Lambda, LambdaStar, PhiBar, PhiStar)
}
\arguments{
\item{parms}{vector of rotation angles that build an orthogonal matrix.}
}
\value{
Output of the \code{optim()} function.
}
\description{
This function uses a heuristic optimization to the
  quadratic loss function of the ex-post approach towards the rotation
  problem.
}

